Initial commit with lots of functionality.
I had already created lots of functionality for this project before
deciding to put it on GitHub.  There is a central window with a
clickable red button.  When you click it, it plays a sound.  The sound
file to be played can be specified in the properties window, along with
repeat options, volume controls, and button label aesthetics.  Save
functionality to come soon.